"Update return types in unarchiver
Changed return types in unarchiver.h and unarchiver.c to be more explicit about data types. This update was done to standardize the return types for extract_file and extract_archive functions from int to ssize_t and uint32_t, respectively, promoting type safety and reducing potential errors".

+++ src/tp2/unarchiver.c
@@ -26,28 +26,28 @@
  *
  * @return The total number of bytes copied, or -1 in case of errors.
  */
-int copy_content(int source, int destination, int size)
+ssize_t copy_content(int source, int destination, ssize_t size)
 {
     char buffer[1024];
-    int total = 0;
+    ssize_t total = 0;
     while (size > 0)
     {
-        int bytes_to_read = ((int) sizeof(buffer) < size) ? (int) sizeof(buffer) : size;
-        int read_bytes = read(source, buffer, bytes_to_read);
+        ssize_t bytes_to_read = ((int) sizeof(buffer) < size) ? (int) sizeof(buffer) : size;
+        ssize_t read_bytes = read(source, buffer, bytes_to_read);
         if (read_bytes < 0)
         {
             perror("Read error in copy_content");
             return -1;
         } else if (read_bytes == 0)
         {
-            fprintf(stderr, "EOF encountered in copy_content with %d bytes left to read\n", size);
+            fprintf(stderr, "EOF encountered in copy_content with %zd bytes left to read\n", size);
             return -1;
         }
 
-        int written_bytes = 0;
+        ssize_t written_bytes = 0;
         while (written_bytes < read_bytes)
         {
-            int result = write(destination, buffer + written_bytes, read_bytes - written_bytes);
+            ssize_t result = write(destination, buffer + written_bytes, read_bytes - written_bytes);
             if (result < 0)
             {
                 perror("Write error in copy_content");
@@ -69,7 +69,7 @@ int copy_content(int source, int destination, int size)
  *
  * @return The total number of bytes written to the extracted file, or -1 in case of errors.
  */
-int extract_file(int fd_archive)
+ssize_t extract_file(int fd_archive)
 {
     uint8_t file_name_size;
     if (read(fd_archive, &file_name_size, sizeof(file_name_size)) != sizeof(file_name_size))
@@ -100,7 +100,7 @@ int extract_file(int fd_archive)
         return -1;
     }
 
-    int res = copy_content(fd_archive, fd_file, file_size);
+    ssize_t res = copy_content(fd_archive, fd_file, file_size);
 
     close(fd_file);
 
@@ -115,7 +115,7 @@ int extract_file(int fd_archive)
  *
  * @return The number of files extracted from the archive, or -1 in case of errors.
  */
-int extract_archive(const char *archive)
+uint32_t extract_archive(const char *archive)
 {
     int fd = open(archive, O_RDONLY);
     if (fd == -1)
@@ -161,7 +161,7 @@ int main(int argc, char **argv)
         return 1;
     }
 
-    int result = extract_archive(argv[1]);
+    uint32_t result = extract_archive(argv[1]);
     if (result == -1)
     {
         perror("Error extracting archive");
